Move code into deprecated directory
