+++ examples/xplatform/deprecated_grpc/echo.proto
