+++ examples/xplatform/deprecated_proto/BUILD
@@ -1,14 +1,11 @@
 load(
-    "@rules_proto//proto:defs.bzl",
+    "@rules_proto//proto:defs.bzl", 
     "proto_library",
-)
-load(
-    "//swift/proto:swift_proto_library.bzl",
-    "swift_proto_library",
 )
 load(
     "//swift:swift.bzl",
     "swift_binary",
+    "swift_proto_library",
 )
 
 licenses(["notice"])
@@ -23,7 +20,7 @@ proto_library(
 
 swift_proto_library(
     name = "example_proto_swift",
-    protos = [":example_proto"],
+    deps = [":example_proto"],
 )
 
 proto_library(
@@ -32,15 +29,13 @@ proto_library(
         "PathToUnderscores/message_1/message.proto",
         "PathToUnderscores/message_2/message.proto",
     ],
+    features = ["swift.generate_path_to_underscores_from_proto_files"],
     strip_import_prefix = "PathToUnderscores",
 )
 
 swift_proto_library(
     name = "example_path_to_underscores_proto_swift",
-    additional_plugin_options = {
-        "FileNaming": "PathToUnderscores",
-    },
-    protos = [":example_path_to_underscores_proto"],
+    deps = [":example_path_to_underscores_proto"],
 )
 
 swift_binary(
