+++ examples/xplatform/deprecated_proto/PathToUnderscores/message_1/message.proto
