+++ examples/xplatform/deprecated_proto/PathToUnderscores/message_2/message.proto
