+++ examples/xplatform/deprecated_proto/example.proto
