+++ examples/xplatform/grpc/service/BUILD
@@ -1,10 +1,10 @@
 load("@rules_proto//proto:defs.bzl", "proto_library")
-load("//swift/proto:swift_proto_library.bzl", "swift_proto_library")
+load("//swift:swift.bzl", swift_proto_library = "new_swift_proto_library")
 
 proto_library(
     name = "service_proto",
     srcs = ["service.proto"],
-    deps = ["//examples/xplatform/new_grpc/service/service_messages:service_messages_proto"],
+    deps = ["//examples/xplatform/grpc/service/service_messages:service_messages_proto"],
 )
 
 swift_proto_library(
@@ -13,7 +13,7 @@ swift_proto_library(
     module_name = "ServiceServer",
     protos = [":service_proto"],
     visibility = ["//visibility:public"],
-    deps = ["//examples/xplatform/new_grpc/service/service_messages:service_messages_swift_proto"],
+    deps = ["//examples/xplatform/grpc/service/service_messages:service_messages_swift_proto"],
 )
 
 swift_proto_library(
@@ -22,7 +22,7 @@ swift_proto_library(
     module_name = "ServiceClient",
     protos = [":service_proto"],
     visibility = ["//visibility:public"],
-    deps = ["//examples/xplatform/new_grpc/service/service_messages:service_messages_swift_proto"],
+    deps = ["//examples/xplatform/grpc/service/service_messages:service_messages_swift_proto"],
 )
 
 swift_proto_library(
@@ -31,13 +31,13 @@ swift_proto_library(
         "ExtraModuleImports": "ServiceClient",
     },
     compiler_deps = [
-        "//examples/xplatform/new_grpc/service:service_client_swift_proto",
+        "//examples/xplatform/grpc/service:service_client_swift_proto",
     ],
     compilers = ["//swift/proto:swift_test_client_grpc"],
     module_name = "ServiceTestClient",
     protos = [":service_proto"],
     visibility = ["//visibility:public"],
     deps = [
-        "//examples/xplatform/new_grpc/service/service_messages:service_messages_swift_proto",
+        "//examples/xplatform/grpc/service/service_messages:service_messages_swift_proto",
     ],
 )
