+++ examples/xplatform/proto/BUILD
@@ -1,8 +1,11 @@
-load("@rules_proto//proto:defs.bzl", "proto_library")
+load(
+    "@rules_proto//proto:defs.bzl",
+    "proto_library",
+)
 load(
     "//swift:swift.bzl",
     "swift_binary",
-    "swift_proto_library",
+    swift_proto_library = "new_swift_proto_library"
 )
 
 licenses(["notice"])
@@ -17,7 +20,7 @@ proto_library(
 
 swift_proto_library(
     name = "example_proto_swift",
-    deps = [":example_proto"],
+    protos = [":example_proto"],
 )
 
 proto_library(
@@ -26,13 +29,15 @@ proto_library(
         "PathToUnderscores/message_1/message.proto",
         "PathToUnderscores/message_2/message.proto",
     ],
-    features = ["swift.generate_path_to_underscores_from_proto_files"],
     strip_import_prefix = "PathToUnderscores",
 )
 
 swift_proto_library(
     name = "example_path_to_underscores_proto_swift",
-    deps = [":example_path_to_underscores_proto"],
+    additional_plugin_options = {
+        "FileNaming": "PathToUnderscores",
+    },
+    protos = [":example_path_to_underscores_proto"],
 )
 
 swift_binary(
