+++ swift/deprecated_proto/proto_gen_utils.bzl
