+++ swift/deprecated_proto/transitions.bzl
